requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pypdf==4.2.0
pinecone-client==3.2.2
openai==1.30.1
streamlit==1.34.0
requests==2.31.0
